[federation] drop experimental reuse fragment optimization
Drop support for the experimental reuse fragment query optimization. This implementation was very slow and buggy due to its complexity. Auto generation of fragments is much simpler algorithm that in most cases produces better results (and much faster).
