+++ apollo-federation/src/operation/optimize.rs
@@ -48,7 +48,6 @@ use super::Containment;
 use super::ContainmentOptions;
 use super::DirectiveList;
 use super::Field;
-use super::FieldSelection;
 use super::Fragment;
 use super::FragmentSpreadSelection;
 use super::HasSelectionKey;
@@ -87,125 +86,6 @@ impl<'a> ReuseContext<'a> {
             fragments,
             operation_variables: Some(operation_variables.iter().map(|var| &var.name).collect()),
         }
-    }
-}
-
-//=============================================================================
-// Add __typename field for abstract types in named fragment definitions
-
-impl NamedFragments {
-    // - Expands all nested fragments
-    // - Applies the provided `mapper` to each selection set of the expanded fragments.
-    // - Finally, re-fragments the nested fragments.
-    // - `mapper` must return a fragment-spread-free selection set.
-    fn map_to_expanded_selection_sets(
-        &self,
-        mut mapper: impl FnMut(&SelectionSet) -> Result<SelectionSet, FederationError>,
-    ) -> Result<NamedFragments, FederationError> {
-        let mut result = NamedFragments::default();
-        // Note: `self.fragments` has insertion order topologically sorted.
-        for fragment in self.fragments.values() {
-            let expanded_selection_set = fragment
-                .selection_set
-                .expand_all_fragments()?
-                .flatten_unnecessary_fragments(
-                    &fragment.type_condition_position,
-                    &Default::default(),
-                    &fragment.schema,
-                )?;
-            let mut mapped_selection_set = mapper(&expanded_selection_set)?;
-            // `mapped_selection_set` must be fragment-spread-free.
-            mapped_selection_set.reuse_fragments(&ReuseContext::for_fragments(&result))?;
-            let updated = Fragment {
-                selection_set: mapped_selection_set,
-                schema: fragment.schema.clone(),
-                name: fragment.name.clone(),
-                type_condition_position: fragment.type_condition_position.clone(),
-                directives: fragment.directives.clone(),
-            };
-            result.insert(updated);
-        }
-        Ok(result)
-    }
-
-    pub(crate) fn add_typename_field_for_abstract_types_in_named_fragments(
-        &self,
-    ) -> Result<Self, FederationError> {
-        // This method is a bit tricky due to potentially nested fragments. More precisely, suppose that
-        // we have:
-        //   fragment MyFragment on T {
-        //     a {
-        //       b {
-        //         ...InnerB
-        //       }
-        //     }
-        //   }
-        //
-        //   fragment InnerB on B {
-        //     __typename
-        //     x
-        //     y
-        //   }
-        // then if we were to "naively" add `__typename`, the first fragment would end up being:
-        //   fragment MyFragment on T {
-        //     a {
-        //       __typename
-        //       b {
-        //         __typename
-        //         ...InnerX
-        //       }
-        //     }
-        //   }
-        // but that's not ideal because the inner-most `__typename` is already within `InnerX`. And that
-        // gets in the way to re-adding fragments (the `SelectionSet::reuse_fragments` method) because if we start
-        // with:
-        //   {
-        //     a {
-        //       __typename
-        //       b {
-        //         __typename
-        //         x
-        //         y
-        //       }
-        //     }
-        //   }
-        // and add `InnerB` first, we get:
-        //   {
-        //     a {
-        //       __typename
-        //       b {
-        //         ...InnerB
-        //       }
-        //     }
-        //   }
-        // and it becomes tricky to recognize the "updated-with-typename" version of `MyFragment` now (we "seem"
-        // to miss a `__typename`).
-        //
-        // Anyway, to avoid this issue, what we do is that for every fragment, we:
-        //  1. expand any nested fragments in its selection.
-        //  2. add `__typename` where we should in that expanded selection.
-        //  3. re-optimize all fragments (using the "updated-with-typename" versions).
-        // which is what `mapToExpandedSelectionSets` gives us.
-
-        if self.is_empty() {
-            // PORT_NOTE: This was an assertion failure in JS version. But, it's actually ok to
-            // return unchanged if empty.
-            return Ok(self.clone());
-        }
-        let updated = self.map_to_expanded_selection_sets(|ss| {
-            // Note: Since `ss` won't have any fragment spreads, `add_typename_field_for_abstract_types`'s return
-            // value won't have any fragment spreads.
-            ss.add_typename_field_for_abstract_types(/*parent_type_if_abstract*/ None)
-        })?;
-        // PORT_NOTE: The JS version asserts if `updated` is empty or not. But, we really want to
-        // check the `updated` has the same set of fragments. To avoid performance hit, only the
-        // size is checked here.
-        if updated.len() != self.len() {
-            return Err(FederationError::internal(
-                "Unexpected change in the number of fragments",
-            ));
-        }
-        Ok(updated)
     }
 }
 
@@ -1264,68 +1144,6 @@ impl NamedFragments {
 //=============================================================================
 // `reuse_fragments` methods (putting everything together)
 
-impl Selection {
-    fn reuse_fragments_inner(
-        &self,
-        context: &ReuseContext<'_>,
-        validator: &mut FieldsConflictMultiBranchValidator,
-        fragments_at_type: &mut FragmentRestrictionAtTypeCache,
-    ) -> Result<Selection, FederationError> {
-        match self {
-            Selection::Field(field) => Ok(field
-                .reuse_fragments_inner(context, validator, fragments_at_type)?
-                .into()),
-            Selection::FragmentSpread(_) => Ok(self.clone()), // Do nothing
-            Selection::InlineFragment(inline_fragment) => Ok(inline_fragment
-                .reuse_fragments_inner(context, validator, fragments_at_type)?
-                .into()),
-        }
-    }
-}
-
-impl FieldSelection {
-    fn reuse_fragments_inner(
-        &self,
-        context: &ReuseContext<'_>,
-        validator: &mut FieldsConflictMultiBranchValidator,
-        fragments_at_type: &mut FragmentRestrictionAtTypeCache,
-    ) -> Result<Self, FederationError> {
-        let Some(base_composite_type): Option<CompositeTypeDefinitionPosition> =
-            self.field.output_base_type()?.try_into().ok()
-        else {
-            return Ok(self.clone());
-        };
-        let Some(ref selection_set) = self.selection_set else {
-            return Ok(self.clone());
-        };
-
-        let mut field_validator = validator.for_field(&self.field);
-
-        // First, see if we can reuse fragments for the selection of this field.
-        let opt = selection_set.try_apply_fragments(
-            &base_composite_type,
-            context,
-            &mut field_validator,
-            fragments_at_type,
-            FullMatchingFragmentCondition::ForFieldSelection,
-        )?;
-
-        let mut optimized = match opt {
-            SelectionSetOrFragment::Fragment(fragment) => {
-                let fragment_selection = FragmentSpreadSelection::from_fragment(
-                    &fragment,
-                    /*directives*/ &Default::default(),
-                );
-                SelectionSet::from_selection(base_composite_type, fragment_selection.into())
-            }
-            SelectionSetOrFragment::SelectionSet(selection_set) => selection_set,
-        };
-        optimized =
-            optimized.reuse_fragments_inner(context, &mut field_validator, fragments_at_type)?;
-        Ok(self.with_updated_selection_set(Some(optimized)))
-    }
-}
-
 /// Return type for `InlineFragmentSelection::reuse_fragments`.
 #[derive(derive_more::From)]
 enum FragmentSelection {
@@ -1343,211 +1161,8 @@ impl From<FragmentSelection> for Selection {
     }
 }
 
-impl InlineFragmentSelection {
-    fn reuse_fragments_inner(
-        &self,
-        context: &ReuseContext<'_>,
-        validator: &mut FieldsConflictMultiBranchValidator,
-        fragments_at_type: &mut FragmentRestrictionAtTypeCache,
-    ) -> Result<FragmentSelection, FederationError> {
-        let optimized;
-
-        let type_condition_position = &self.inline_fragment.type_condition_position;
-        if let Some(type_condition_position) = type_condition_position {
-            let opt = self.selection_set.try_apply_fragments(
-                type_condition_position,
-                context,
-                validator,
-                fragments_at_type,
-                FullMatchingFragmentCondition::ForInlineFragmentSelection {
-                    type_condition_position,
-                    directives: &self.inline_fragment.directives,
-                },
-            )?;
-
-            match opt {
-                SelectionSetOrFragment::Fragment(fragment) => {
-                    // We're fully matching the sub-selection. If the fragment condition is also
-                    // this element condition, then we can replace the whole element by the spread
-                    // (not just the sub-selection).
-                    if *type_condition_position == fragment.type_condition_position {
-                        // Optimized as `...<fragment>`, dropping the original inline spread (`self`).
-
-                        // Note that `FullMatchingFragmentCondition::ForInlineFragmentSelection`
-                        // above guarantees that this element directives are a superset of the
-                        // fragment directives. But there can be additional directives, and in that
-                        // case they should be kept on the spread.
-                        // PORT_NOTE: We are assuming directives on fragment definitions are
-                        //            carried over to their spread sites as JS version does, which
-                        //            is handled differently in Rust version (see `FragmentSpread`).
-                        let directives: executable::DirectiveList = self
-                            .inline_fragment
-                            .directives
-                            .iter()
-                            .filter(|d1| !fragment.directives.iter().any(|d2| *d1 == d2))
-                            .cloned()
-                            .collect();
-                        return Ok(
-                            FragmentSpreadSelection::from_fragment(&fragment, &directives).into(),
-                        );
-                    } else {
-                        // Otherwise, we keep this element and use a sub-selection with just the spread.
-                        // Optimized as `...on <type_condition_position> { ...<fragment> }`
-                        optimized = SelectionSet::from_selection(
-                            type_condition_position.clone(),
-                            FragmentSpreadSelection::from_fragment(
-                                &fragment,
-                                /*directives*/ &Default::default(),
-                            )
-                            .into(),
-                        );
-                    }
-                }
-                SelectionSetOrFragment::SelectionSet(selection_set) => {
-                    optimized = selection_set;
-                }
-            }
-        } else {
-            optimized = self.selection_set.clone();
-        }
-
-        Ok(InlineFragmentSelection::new(
-            self.inline_fragment.clone(),
-            // Then, recurse inside the field sub-selection (note that if we matched some fragments
-            // above, this recursion will "ignore" those as `FragmentSpreadSelection`'s
-            // `reuse_fragments()` is a no-op).
-            optimized.reuse_fragments_inner(context, validator, fragments_at_type)?,
-        )
-        .into())
-    }
-}
-
-impl SelectionSet {
-    fn reuse_fragments_inner(
-        &self,
-        context: &ReuseContext<'_>,
-        validator: &mut FieldsConflictMultiBranchValidator,
-        fragments_at_type: &mut FragmentRestrictionAtTypeCache,
-    ) -> Result<SelectionSet, FederationError> {
-        self.lazy_map(context.fragments, |selection| {
-            Ok(selection
-                .reuse_fragments_inner(context, validator, fragments_at_type)?
-                .into())
-        })
-    }
-
-    fn contains_fragment_spread(&self) -> bool {
-        self.iter().any(|selection| {
-            matches!(selection, Selection::FragmentSpread(_))
-                || selection
-                    .selection_set()
-                    .map(|subselection| subselection.contains_fragment_spread())
-                    .unwrap_or(false)
-        })
-    }
-
-    /// ## Errors
-    /// Returns an error if the selection set contains a named fragment spread.
-    fn reuse_fragments(&mut self, context: &ReuseContext<'_>) -> Result<(), FederationError> {
-        if context.fragments.is_empty() {
-            return Ok(());
-        }
-
-        if self.contains_fragment_spread() {
-            return Err(FederationError::internal("reuse_fragments() must only be used on selection sets that do not contain named fragment spreads"));
-        }
-
-        // Calling reuse_fragments() will not match a fragment that would have expanded at
-        // top-level. That is, say we have the selection set `{ x y }` for a top-level `Query`, and
-        // we have a fragment
-        // ```
-        // fragment F on Query {
-        //   x
-        //   y
-        // }
-        // ```
-        // then calling `self.reuse_fragments(fragments)` would only apply check if F apply to
-        // `x` and then `y`.
-        //
-        // To ensure the fragment match in this case, we "wrap" the selection into a trivial
-        // fragment of the selection parent, so in the example above, we create selection `... on
-        // Query { x y }`. With that, `reuse_fragments` will correctly match on the `on Query`
-        // fragment; after which we can unpack the final result.
-        let wrapped = InlineFragmentSelection::from_selection_set(
-            self.type_position.clone(), // parent type
-            self.clone(),               // selection set
-            Default::default(),         // directives
-        );
-        let mut validator = FieldsConflictMultiBranchValidator::from_initial_validator(
-            FieldsConflictValidator::from_selection_set(self),
-        );
-        let optimized = wrapped.reuse_fragments_inner(
-            context,
-            &mut validator,
-            &mut FragmentRestrictionAtTypeCache::default(),
-        )?;
-
-        // Now, it's possible we matched a full fragment, in which case `optimized` will be just
-        // the named fragment, and in that case we return a singleton selection with just that.
-        // Otherwise, it's our wrapping inline fragment with the sub-selections optimized, and we
-        // just return that subselection.
-        *self = match optimized {
-            FragmentSelection::FragmentSpread(spread) => {
-                SelectionSet::from_selection(self.type_position.clone(), spread.into())
-            }
-            FragmentSelection::InlineFragment(inline_fragment) => inline_fragment.selection_set,
-        };
-        Ok(())
-    }
-}
 
 impl Operation {
-    // PORT_NOTE: The JS version of `reuse_fragments` takes an optional `minUsagesToOptimize` argument.
-    //            However, it's only used in tests. So, it's removed in the Rust version.
-    const DEFAULT_MIN_USAGES_TO_OPTIMIZE: u32 = 2;
-
-    // `fragments` - rebased fragment definitions for the operation's subgraph
-    // - `self.selection_set` must be fragment-spread-free.
-    fn reuse_fragments_inner(
-        &mut self,
-        fragments: &NamedFragments,
-        min_usages_to_optimize: u32,
-    ) -> Result<(), FederationError> {
-        if fragments.is_empty() {
-            return Ok(());
-        }
-
-        // Optimize the operation's selection set by re-using existing fragments.
-        let before_optimization = self.selection_set.clone();
-        self.selection_set
-            .reuse_fragments(&ReuseContext::for_operation(fragments, &self.variables))?;
-        if before_optimization == self.selection_set {
-            return Ok(());
-        }
-
-        // Optimize the named fragment definitions by dropping low-usage ones.
-        let mut final_fragments = fragments.clone();
-        let final_selection_set =
-            final_fragments.reduce(&self.selection_set, min_usages_to_optimize)?;
-
-        self.selection_set = final_selection_set;
-        self.named_fragments = final_fragments;
-        Ok(())
-    }
-
-    /// Optimize the parsed size of the operation by applying fragment spreads. Fragment spreads
-    /// are reused from the original user-provided fragments.
-    ///
-    /// `fragments` - rebased fragment definitions for the operation's subgraph
-    ///
-    // PORT_NOTE: In JS, this function was called "optimize".
-    pub(crate) fn reuse_fragments(
-        &mut self,
-        fragments: &NamedFragments,
-    ) -> Result<(), FederationError> {
-        self.reuse_fragments_inner(fragments, Self::DEFAULT_MIN_USAGES_TO_OPTIMIZE)
-    }
-
     /// Optimize the parsed size of the operation by generating fragments based on the selections
     /// in the operation.
     pub(crate) fn generate_fragments(&mut self) -> Result<(), FederationError> {
@@ -1565,16 +1180,6 @@ impl Operation {
         generator.visit_selection_set(&mut self.selection_set)?;
         self.named_fragments = generator.into_inner();
         Ok(())
-    }
-
-    /// Used by legacy roundtrip tests.
-    /// - This lowers `min_usages_to_optimize` to `1` in order to make it easier to write unit tests.
-    #[cfg(test)]
-    fn reuse_fragments_for_roundtrip_test(
-        &mut self,
-        fragments: &NamedFragments,
-    ) -> Result<(), FederationError> {
-        self.reuse_fragments_inner(fragments, /*min_usages_to_optimize*/ 1)
     }
 
     // PORT_NOTE: This mirrors the JS version's `Operation.expandAllFragments`. But this method is
