+++ apollo-federation/tests/query_plan/build_query_plan_tests/fragment_autogeneration.rs
@@ -1,5 +1,12 @@
 use apollo_federation::query_plan::query_planner::QueryPlannerConfig;
 
+fn generate_fragments_config() -> QueryPlannerConfig {
+    QueryPlannerConfig {
+        generate_query_fragments: true,
+        ..Default::default()
+    }
+}
+
 const SUBGRAPH: &str = r#"
       directive @custom on INLINE_FRAGMENT | FRAGMENT_SPREAD
 
@@ -25,7 +32,7 @@ const SUBGRAPH: &str = r#"
 #[test]
 fn it_respects_generate_query_fragments_option() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: SUBGRAPH,
     );
     assert_plan!(
@@ -73,7 +80,7 @@ fn it_respects_generate_query_fragments_option() {
 #[test]
 fn it_handles_nested_fragment_generation() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: SUBGRAPH,
     );
     assert_plan!(
@@ -131,10 +138,250 @@ fn it_handles_nested_fragment_generation() {
     );
 }
 
+// TODO this test shows a clearly worse plan than reused fragments when fragments
+// target concrete types
+#[test]
+fn it_handles_nested_fragment_generation_from_operation_with_fragments() {
+    let planner = planner!(
+        config = generate_fragments_config(),
+        Subgraph1: r#"
+          type Query {
+            a: Anything
+          }
+
+          union Anything = A1 | A2 | A3
+
+          interface Foo {
+            foo: String
+            child: Foo
+            child2: Foo
+          }
+
+          type A1 implements Foo {
+            foo: String
+            child: Foo
+            child2: Foo
+          }
+
+          type A2 implements Foo {
+            foo: String
+            child: Foo
+            child2: Foo
+          }
+
+          type A3 implements Foo {
+            foo: String
+            child: Foo
+            child2: Foo
+          }
+        "#,
+    );
+    assert_plan!(
+        &planner,
+        r#"
+          query {
+            a {
+              ... on A1 {
+                ...FooSelect
+              }
+              ... on A2 {
+                ...FooSelect
+              }
+              ... on A3 {
+                ...FooSelect
+              }
+            }
+          }
+
+          fragment FooSelect on Foo {
+            __typename
+            foo
+            child {
+              ...FooChildSelect
+            }
+            child2 {
+              ...FooChildSelect
+            }
+          }
+
+          fragment FooChildSelect on Foo {
+            __typename
+            foo
+            child {
+              child {
+                child {
+                  foo
+                }
+              }
+            }
+          }
+        "#,
+
+        // This is a test case that shows worse result
+        // QueryPlan {
+        //           Fetch(service: "Subgraph1") {
+        //             {
+        //               a {
+        //                 __typename
+        //                 ... on A1 {
+        //                   ...FooSelect
+        //                 }
+        //                 ... on A2 {
+        //                   ...FooSelect
+        //                 }
+        //                 ... on A3 {
+        //                   ...FooSelect
+        //                 }
+        //               }
+        //             }
+        //
+        //             fragment FooChildSelect on Foo {
+        //               __typename
+        //               foo
+        //               child {
+        //                 __typename
+        //                 child {
+        //                   __typename
+        //                   child {
+        //                     __typename
+        //                     foo
+        //                   }
+        //                 }
+        //               }
+        //             }
+        //
+        //             fragment FooSelect on Foo {
+        //               __typename
+        //               foo
+        //               child {
+        //                 ...FooChildSelect
+        //               }
+        //               child2 {
+        //                 ...FooChildSelect
+        //               }
+        //             }
+        //           },
+        //         }
+        @r###"
+    QueryPlan {
+      Fetch(service: "Subgraph1") {
+        {
+          a {
+            __typename
+            ..._generated_onA14_0
+            ..._generated_onA24_0
+            ..._generated_onA34_0
+          }
+        }
+
+        fragment _generated_onA14_0 on A1 {
+          __typename
+          foo
+          child {
+            __typename
+            foo
+            child {
+              __typename
+              child {
+                __typename
+                child {
+                  __typename
+                  foo
+                }
+              }
+            }
+          }
+          child2 {
+            __typename
+            foo
+            child {
+              __typename
+              child {
+                __typename
+                child {
+                  __typename
+                  foo
+                }
+              }
+            }
+          }
+        }
+
+        fragment _generated_onA24_0 on A2 {
+          __typename
+          foo
+          child {
+            __typename
+            foo
+            child {
+              __typename
+              child {
+                __typename
+                child {
+                  __typename
+                  foo
+                }
+              }
+            }
+          }
+          child2 {
+            __typename
+            foo
+            child {
+              __typename
+              child {
+                __typename
+                child {
+                  __typename
+                  foo
+                }
+              }
+            }
+          }
+        }
+
+        fragment _generated_onA34_0 on A3 {
+          __typename
+          foo
+          child {
+            __typename
+            foo
+            child {
+              __typename
+              child {
+                __typename
+                child {
+                  __typename
+                  foo
+                }
+              }
+            }
+          }
+          child2 {
+            __typename
+            foo
+            child {
+              __typename
+              child {
+                __typename
+                child {
+                  __typename
+                  foo
+                }
+              }
+            }
+          }
+        }
+      },
+    }
+    "###
+    );
+}
+
 #[test]
 fn it_handles_fragments_with_one_non_leaf_field() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: SUBGRAPH,
     );
 
@@ -183,7 +430,7 @@ fn it_handles_fragments_with_one_non_leaf_field() {
 #[test]
 fn it_migrates_skip_include() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: SUBGRAPH,
     );
     assert_plan!(
@@ -250,10 +497,11 @@ fn it_migrates_skip_include() {
     "###
     );
 }
+
 #[test]
 fn it_identifies_and_reuses_equivalent_fragments_that_arent_identical() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: SUBGRAPH,
     );
     assert_plan!(
@@ -301,7 +549,7 @@ fn it_identifies_and_reuses_equivalent_fragments_that_arent_identical() {
 #[test]
 fn fragments_that_share_a_hash_but_are_not_identical_generate_their_own_fragment_definitions() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: SUBGRAPH,
     );
     assert_plan!(
@@ -354,7 +602,7 @@ fn fragments_that_share_a_hash_but_are_not_identical_generate_their_own_fragment
 #[test]
 fn same_as_js_router798() {
     let planner = planner!(
-        config = QueryPlannerConfig { generate_query_fragments: true, reuse_query_fragments: false, ..Default::default() },
+        config = generate_fragments_config(),
         Subgraph1: r#"
             interface Interface { a: Int }
             type Y implements Interface { a: Int b: Int }
@@ -398,6 +646,7 @@ fn same_as_js_router798() {
 #[test]
 fn works_with_key_chains() {
     let planner = planner!(
+        config = generate_fragments_config(),
         Subgraph1: r#"
       type Query {
         t: T
@@ -471,15 +720,206 @@ fn works_with_key_chains() {
               }
             } =>
             {
-              ... on T {
-                x
-                y
-              }
+              ..._generated_onT2_0
+            }
+
+            fragment _generated_onT2_0 on T {
+              x
+              y
             }
           },
         },
       },
     }
   "###
+    );
+}
+
+#[test]
+fn another_mix_of_fragments_indirection_and_unions() {
+    // This tests that the issue reported on https://github.com/apollographql/router/issues/3172 is resolved.
+    let planner = planner!(
+        config = generate_fragments_config(),
+        Subgraph1: r#"
+          type Query {
+            owner: Owner!
+          }
+
+          interface OItf {
+            id: ID!
+            v0: String!
+          }
+
+          type Owner implements OItf {
+            id: ID!
+            v0: String!
+            u: [U]
+          }
+
+          union U = T1 | T2
+
+          interface I {
+            id1: ID!
+            id2: ID!
+          }
+
+          type T1 implements I {
+            id1: ID!
+            id2: ID!
+            owner: Owner!
+          }
+
+          type T2 implements I {
+            id1: ID!
+            id2: ID!
+          }
+        "#,
+    );
+    assert_plan!(
+        &planner,
+        r#"
+          {
+            owner {
+              u {
+                ... on I {
+                  id1
+                  id2
+                }
+                ...Fragment1
+                ...Fragment2
+              }
+            }
+          }
+
+          fragment Fragment1 on T1 {
+            owner {
+              ... on Owner {
+                ...Fragment3
+              }
+            }
+          }
+
+          fragment Fragment2 on T2 {
+            ...Fragment4
+            id1
+          }
+
+          fragment Fragment3 on OItf {
+            v0
+          }
+
+          fragment Fragment4 on I {
+            id1
+            id2
+            __typename
+          }
+        "#,
+        @r###"
+        QueryPlan {
+          Fetch(service: "Subgraph1") {
+            {
+              owner {
+                u {
+                  __typename
+                  ..._generated_onI3_0
+                  ..._generated_onT11_0
+                  ..._generated_onT23_0
+                }
+              }
+            }
+
+            fragment _generated_onI3_0 on I {
+              __typename
+              id1
+              id2
+            }
+
+            fragment _generated_onT11_0 on T1 {
+              owner {
+                v0
+              }
+            }
+
+            fragment _generated_onT23_0 on T2 {
+              __typename
+              id1
+              id2
+            }
+          },
+        }
+      "###
+    );
+
+    assert_plan!(
+        &planner,
+        r#"
+          {
+            owner {
+              u {
+                ... on I {
+                  id1
+                  id2
+                }
+                ...Fragment1
+                ...Fragment2
+              }
+            }
+          }
+
+          fragment Fragment1 on T1 {
+            owner {
+              ... on Owner {
+                ...Fragment3
+              }
+            }
+          }
+
+          fragment Fragment2 on T2 {
+            ...Fragment4
+            id1
+          }
+
+          fragment Fragment3 on OItf {
+            v0
+          }
+
+          fragment Fragment4 on I {
+            id1
+            id2
+          }
+        "#,
+        @r###"
+        QueryPlan {
+          Fetch(service: "Subgraph1") {
+            {
+              owner {
+                u {
+                  __typename
+                  ..._generated_onI3_0
+                  ..._generated_onT11_0
+                  ..._generated_onT22_0
+                }
+              }
+            }
+
+            fragment _generated_onI3_0 on I {
+              __typename
+              id1
+              id2
+            }
+
+            fragment _generated_onT11_0 on T1 {
+              owner {
+                v0
+              }
+            }
+
+            fragment _generated_onT22_0 on T2 {
+              id1
+              id2
+            }
+          },
+        }
+      "###
     );
 }
